Implement basic lobby and rooms
Lobby is the meeting place for users that wants to collaborate. It
resembles sugar's neighborhood. Users can publish their presence
and their shared activities.

Rooms are private communication channels. Each "room" provides an
exclusive channel for a particular activity instance and the users
participating in the shared activity.

This version is implemented using socket.io. At this point I am not
sure I am using it in the "right" way, but it is sufficient to get
our prototype working.

Signed-off-by: Martin Abente Lahaye <tch@sugarlabs.org>